castle_inside, new bowser portrait and fixed placeholder wf painting
bowser: 12.rgba16, 13.rgba16, 14.rgba16, 15.rgba16, bowser_rlt_rlt
whomp's fortress painting (placeholder): 21.rgba16, 22.rgba16,

the whomp's fortress painting would display a white horizontal line through the middle. that's because the texture was designed in a way that worked around the issue of one texture appearing to cutoff as it connects to another when viewed ingame; this applied to the painting in the whomp's fortress room but not the painting on the second floor.