textures/outside, improved hill/bridge/statue "noise" texture (04800)
this texture can be seen on the sides of the moat bridge and on the hills in front of the castle, as well as the star statue in the courtyard. texture was really noisy before so it's been made less noisy so to not appear jarring when seen next to the other textures in our pack. in general, it also better resembles the original texture now.