pls read commit description
this branch as the name implies aims to downscale the render 96 textures back to n64 standards, as it stands now  only a small sample of textures have been dowscaled. we hope to downscale more of these in the future.

read the credits to know the people involved in the downscaling